Changing parameters to FromFrom - changing type float to decimal

+++ Server/Butterfly.Services/Interfaces/ICalculatorService.cs
@@ -2,9 +2,9 @@
 {
     public interface ICalculatorService
     {
-        float Add(float number1, float number2);
-        float Subtract(float number1, float number2);
-        float Divide(float number1, float number2);
-        float Multiply(float number1, float number2);
+        decimal Add(decimal number1, decimal number2);
+        decimal Subtract(decimal number1, decimal number2);
+        decimal Divide(decimal number1, decimal number2);
+        decimal Multiply(decimal number1, decimal number2);
     }
 }
